// src/app/sidur_avoda/shared/types/tavlaot-av.ts
export  class TavlaotAv {
    id_premia_tavlat_av: number;
    table_id: number;
    teur_premia: string;
    premia: number;
    kamut?: number;
    kod_mesima?: number;
    meshtamesh_medaveach: number;
    sw_pail: boolean;
    sw_male: boolean;
    simochin?: string;
    id_zman: number;
    heara: string;
    deleted: number;
    user_name?: string;
    moed_idkun?: Date;
    zman_hatchala?: Date;
    zman_sof: Date;
    full_name: string;
    teur_sug_premia_tavlot_av: string;
    id_sug_mesima: number;
    k_sug_mesima: string;
    teur_sug_mesima: string;
    k_user: number;
    premia_user_full_name: string;
    mispar_nahag: number;
    constructor() { }
}